Fix palette timing

+++ src/huc6260.h
@@ -98,7 +98,7 @@ private:
     u16* m_color_table;
     u8* m_frame_buffer;
     u8* m_scale_buffer;
-    u32* m_vce_buffer;
+    u16* m_vce_buffer;
     s32* m_line_speed;
     bool m_multiple_speeds;
     bool m_scaled_width;
